FEC-9909 Added a KavaHelper class to return the plugin config, because it is common for both OTT and OVP. Updated code accordingly. Added the Common folder to the OTT subspec.
